requirements_CapsNet.txt
absl-py==0.9.0
astor==0.8.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
configparser==5.0.0
cycler==0.10.0
decorator==4.4.2
docker-pycreds==0.4.0
gast==0.2.2
gitdb==4.0.5
GitPython==3.1.2
google-pasta==0.2.0
gql==0.2.0
graphql-core==1.1
grpcio==1.28.1
h5py==2.10.0
idna==2.9
imageio==2.8.0
joblib==0.14.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.2.1
networkx==2.4
numpy==1.18.4
nvidia-ml-py3==7.352.0
opt-einsum==3.2.1
pandas==1.0.3
pathtools==0.1.2
Pillow==7.1.2
promise==2.3
protobuf==3.11.3
psutil==5.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
sentry-sdk==0.14.3
shortuuid==1.0.1
six==1.14.0
smmap==3.0.4
subprocess32==3.5.4
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
Theano==1.0.4
urllib3==1.25.9
wandb==0.8.35
watchdog==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
